LayoutTests/platform/wpe/imported/w3c/web-platform-tests/html/semantics/document-metadata/the-meta-element/color-scheme/meta-color-scheme-remove-expected.txt


FAIL First meta applies. assert_equals: expected (string) "dark" but got (undefined) undefined
FAIL Second meta applies after first one is removed. assert_equals: expected (string) "light" but got (undefined) undefined
FAIL Initial color-scheme with both meta elements removed. assert_equals: expected (string) "normal" but got (undefined) undefined
